requirements.txt
Flask==3.0.3
gunicorn==21.2.0
python-dotenv==1.0.0
cloudinary==1.44.1
Pillow==10.0.0